#20 Overhaul Message
Message was missing various fields that it would normally have access to when its
supposed to represent most/all fields for an email.
Those fields were msgBcc and msgReplyTo.
However, Message could also be turned into a utility to be able to send on its own. For this purpose, msgSender was added on top.

Further, emails can be represented as "name" <email>
This can only be done via an object type that can include a name.